Rename spark3 integration tests
