+++ integration-test/spark3/src/test/java/org/apache/uniffle/test/AQERepartitionIT.java
@@ -39,7 +39,7 @@ import org.apache.uniffle.storage.util.StorageType;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class AQERepartitionTest extends SparkIntegrationTestBase {
+public class AQERepartitionIT extends SparkIntegrationTestBase {
 
   @BeforeAll
   public static void setupServers() throws Exception {
