+++ integration-test/spark3/src/test/java/org/apache/uniffle/test/GetReaderIT.java
@@ -62,7 +62,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class GetReaderTest extends IntegrationTestBase {
+public class GetReaderIT extends IntegrationTestBase {
 
   @Test
   public void test() throws Exception {
